backend/requirements.txt
# ==================== Django 核心 ====================
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# ==================== 數據庫 ====================
psycopg2-binary==2.9.7
redis==4.6.0

# ==================== 認證與安全 ====================
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
django-oauth-toolkit==1.7.1
cryptography==41.0.7

# ==================== 文件處理 ====================
Pillow==10.0.1
django-storages==1.14.2
boto3==1.29.7

# ==================== 程式碼高亮 ====================
Pygments==2.16.1

# ==================== 搜索引擎 ====================
elasticsearch==8.10.0
elasticsearch-dsl==8.9.0

# ==================== 緩存與會話 ====================
django-redis==5.4.0

# ==================== 異步任務 ====================
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# ==================== WebSocket ====================
channels==4.0.0
channels-redis==4.1.0

# ==================== API 文檔 ====================
drf-spectacular==0.26.5

# ==================== 開發工具 ====================
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# ==================== 監控與日誌 ====================
sentry-sdk==1.38.0
django-loguru==1.0.0

# ==================== 郵件服務 ====================
django-anymail==10.2

# ==================== 工具庫 ====================
python-decouple==3.8
requests==2.31.0
python-dateutil==2.8.2
shortuuid==1.0.11

# ==================== 測試 ====================
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# ==================== 開發環境 ====================
black==23.9.1
isort==5.12.0
flake8==6.1.0 